+++ package.json
@@ -8,7 +8,7 @@
   },
   "scripts": {
     "prettier:frontend": "prettier ./frontend/src --write",
-    "prettier:server": "prettier ./server --write",
+    "prettier:server": "prettier ./server/**/*.ts --write",
     "dev": "docker-compose --env-file ./.env -f docker-compose.dev.yml up --build",
     "clear-dev-data": "docker-compose -f docker-compose.dev.yml down && docker volume rm dd_database && rm -rf ./redis_data",
     "prepare": "husky install"
